section: change semantics to conform with other implementations
The previous implementation would throw an exception if the section did
not correspond to an object, array, or boolean. This eliminated the
possibility of performing truthy checks on the keys of the current
context, and proceeding accordingly.

Now, if the key corresponds to a truthy value that is not an object or
array, the test will succeed, using the current context as the context
for the section. This is in line with the behavior of other impls,
namely http://trymustache.com.