strict: add less strict option for template checks
The code previously would fail with an error if a template check
received an object of the wrong type. The code still does that, but has
an optional argument to control the behavior. If ~strict:true, which is
the default, then the code exhibits the old behavior. If ~strict:false
then "type errors" will result in taking the false branch.