fold: properly define fold function
A fold for a type t should take a function for each constructor. The
function is passed all the arguments of the corresponding constructor.
If the constructor contains a value of type t, the "folded" t should be
passed to the functions. A fold should only take an initial accumulator
of it has a constructor that takes no arguments, like [] in the case of
lists.