Handle "standalone" elements
For example, " {{!x}} \n" is a standalone comment (it is alone on its line,
ignoring the whitespace), so the whitespace around it (including the newline)
must be removed.

We implement that by wrapping the existing unmodified lexer with hand-written
code that performs the removing of RAW tokens before and after a standalone
element. The wrapped lexer is then fed to the unmodified parser.