mustache_lexer: raise a proper Error instead of Failure
An input such as "{{" would previously fail with an error from the
lexing runtime:

> Fatal error: exception Failure("lexing: empty token")

The reason is that the lexer is written in such a way that it forces
the lexing of input in places where no input may be available.

A long-term fix would be to write the lexer in a different, more
standard style, where we don't try to "bundle" different information
in a single tokan.

The current fix is simply to make all tokens nullable by raising an
explicit Error exception when reaching end-of-file.