have partials interpreted as template inclusion in mustache_cli
(Before this PR, any use of partials will fail in the command-line
tool.)

Partial names are turned into filenames by adding ".mustache" and then
resolveas paths relative to the current working directory (from which
the tool is invoked, not the directory where the template file
is located). For example, `{{>foo/bar}}` would include
"foo/bar.mustache".

This behavior is used by other command-line mustache processors,
typically the Ruby implementation (`gem install mustache`), so it is
somewhat-standard for a default behavior of partials.

Note: the Ruby implementation also supports {{> ../foo/bar}}, which is
not accepted by our current lexer (partial names must use dots in
a way that would be valid for other tags). The lexer could be relaxed
to allow those, but this can come as a later change.