reimplement and document the handling of standalone tokens
The previous approach to standalone-token handling would peek into
a prefix of the line, and decide what to do based on the prefix. This
relied on the property that we would have at most one standalone token
per line. This is unfortunately not true, consider:

```
Begin.
{{#foo}} {{#bar}}
Middle.
{{/bar}} {{/foo}}
End.
```

The new approach processes the whole line at once, failing if it
encounters non-whitespace non-standalone token.