A compatibility test against the ocaml-mustache v2.0.0 API
This test is designed to get a sense of the backward-compatibility
impact of changes to the ocaml-mustache library.

mustache_v200.mli is exactly a copy of the Mustache interface as it
existed in the version v2.0.0, and it should not change.

mustache_v200.ml is a reimplementation of this interface using the
current ocaml-mustache library. If the library changes, this
reimplementation will have to be fixed as well; the invasiveness of
this fix can be used to estimate the invasiveness of the change for
end-users.

user_program.ml is written against this interface, and we check that
it keeps working.

v2.0.0 was chosen because it does not contain an explicit AST
definition or "fold" function -- whose compatibility breaks for
basically any new language feature or representation
change (for example when adding locations in the AST).